requirements.txt
cloudpickle==3.1.2
colorama==0.4.6
imbalanced-learn==0.14.1
imblearn==0.0
joblib==1.5.3
llvmlite==0.46.0
numba==0.64.0
numpy==2.4.1
packaging==26.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.17.0
shap==0.51.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
tabulate==0.10.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
xgboost==3.1.3
